+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - #1325 Refactored CacheMachine.h and CacheMachine.cpp 
 - #1322 Updated and enabled several E2E tests
 - #1333 Fixing build due to cudf update
+- #1344 Removed GPUCacheDataMetadata class 
 
 ## Bug Fixes
 - #1335 Fixing uninitialized var in orc metadata and handling the parseMetadata exceptions properly